Added second queue, and queue selection probability. Need to configure if statements to accnt for new q's.

+++ settings.h
@@ -7,7 +7,7 @@ public:
     int lambda;
     int mu;
     size_t queue_sz;
-    int phi;
+    double phi;
 private:
 };
